[i18n] move i18n servive to @shared
