+++ ui/src/app/@shared/config/i18nSupportList.ts
