+++ ui/src/app/@shared/service/i18n/index.ts
@@ -1,0 +1,2 @@
+export * from './i18n.service';
+export * from '@tide-config/i18nSupportList';
