--- ui/src/app/config/i18n/index.ts
@@ -1,2 +1,0 @@
-export * from './i18n.service';
-export * from './i18nSupportList';
